fix: Added email notifications to producer for activate and suspend (PIN-10348)
