--- packages/email-notification-dispatcher/src/handlers/eservices/handleEserviceDescriptorArchivedToConsumer.ts
@@ -1,115 +1,0 @@
-import {
-  Agreement,
-  DescriptorId,
-  EmailNotificationMessagePayload,
-  fromEServiceV2,
-  generateId,
-  missingKafkaMessageDataError,
-  NotificationType,
-  unsafeBrandId,
-} from "pagopa-interop-models";
-import {
-  eventMailTemplateType,
-  getRecipientsForTenants,
-  mapRecipientToEmailPayload,
-  retrieveDescriptor,
-  retrieveHTMLTemplate,
-  retrieveTenant,
-} from "pagopa-interop-notification-commons";
-import { EServiceDescriptorHandlerParams } from "../../models/handlerParams.js";
-import { config } from "../../config/config.js";
-
-const notificationType: NotificationType = "eserviceStateChangedToConsumer";
-
-export async function handleEserviceDescriptorArchivedToConsumer(
-  data: EServiceDescriptorHandlerParams
-): Promise<EmailNotificationMessagePayload[]> {
-  const {
-    eserviceV2Msg,
-    descriptorId: descriptorIdFromEvent,
-    readModelService,
-    logger,
-    templateService,
-    correlationId,
-  } = data;
-
-  if (!eserviceV2Msg) {
-    throw missingKafkaMessageDataError(
-      "eservice",
-      "EServiceDescriptorArchived"
-    );
-  }
-
-  const eservice = fromEServiceV2(eserviceV2Msg);
-  const descriptorId = unsafeBrandId<DescriptorId>(descriptorIdFromEvent);
-  const descriptor = retrieveDescriptor(eservice, descriptorId);
-
-  // Discriminator: skip auto-archive routine (Deprecated/Suspended -> Archived)
-  if (!descriptor.archivingSchedule) {
-    logger.info(
-      `Skipping email notification for "EServiceDescriptorArchived" without archivingSchedule (eservice ${eservice.id}, descriptor ${descriptor.id}) — routine auto-archiving`
-    );
-    return [];
-  }
-
-  const [htmlTemplate, producer, agreements] = await Promise.all([
-    retrieveHTMLTemplate(
-      eventMailTemplateType.eserviceArchivingDescriptorArchivedToConsumerMailTemplate
-    ),
-    retrieveTenant(eservice.producerId, readModelService),
-    readModelService.getAgreementsByEserviceId(eservice.id, {
-      includeArchived: true,
-    }),
-  ]);
-
-  if (!agreements || agreements.length === 0) {
-    return [];
-  }
-
-  const tenants = await readModelService.getTenantsById(
-    agreements.map((a: Agreement) => a.consumerId)
-  );
-
-  const targets = await getRecipientsForTenants({
-    tenants,
-    notificationType,
-    readModelService,
-    logger,
-    includeTenantContactEmails: false,
-  });
-
-  if (targets.length === 0) {
-    return [];
-  }
-
-  const subject = `Archiviazione anticipata della versione ${descriptor.version} dell'e-service "${eservice.name}"`;
-
-  return targets.flatMap((t) => {
-    const tenant = tenants.find((x) => x.id === t.tenantId);
-    if (!tenant) {
-      return [];
-    }
-    return [
-      {
-        correlationId: correlationId ?? generateId(),
-        email: {
-          subject,
-          body: templateService.compileHtml(htmlTemplate, {
-            title: subject,
-            notificationType,
-            entityId: `${eservice.id}/${descriptor.id}`,
-            ...(t.type === "Tenant" ? { recipientName: tenant.name } : {}),
-            eserviceName: eservice.name,
-            eserviceVersion: descriptor.version,
-            producerName: producer.name,
-            ctaLabel: `Visualizza e-service`,
-            selfcareId: t.selfcareId,
-            bffUrl: config.bffUrl,
-          }),
-        },
-        tenantId: t.tenantId,
-        ...mapRecipientToEmailPayload(t),
-      },
-    ];
-  });
-}
