+++ packages/email-notification-dispatcher/src/handlers/eservices/handleEserviceEvent.ts
@@ -8,8 +8,10 @@ import { handleEserviceDescriptorPublished } from "./handleEserviceDescriptorPub
 import { handleEserviceDescriptorSubmittedByDelegate } from "./handleEserviceDescriptorSubmittedByDelegate.js";
 import { handleEserviceDescriptorApprovedByDelegator } from "./handleEserviceDescriptorApprovedByDelegator.js";
 import { handleEserviceDescriptorRejectedByDelegator } from "./handleEserviceDescriptorRejectedByDelegator.js";
-import { handleEserviceDescriptorActivated } from "./handleEserviceDescriptorActivated.js";
-import { handleEserviceDescriptorSuspended } from "./handleEserviceDescriptorSuspended.js";
+import { handleEserviceDescriptorActivatedToConsumer } from "./handleEserviceDescriptorActivatedToConsumer.js";
+import { handleEserviceDescriptorActivatedToProducer } from "./handleEserviceDescriptorActivatedToProducer.js";
+import { handleEserviceDescriptorSuspendedToConsumer } from "./handleEserviceDescriptorSuspendedToConsumer.js";
+import { handleEserviceDescriptorSuspendedToProducer } from "./handleEserviceDescriptorSuspendedToProducer.js";
 import { handleEserviceStateChanged } from "./handleEserviceStateChanged.js";
 import { handleEserviceDescriptorArchivingScheduledToProducer } from "./handleEserviceDescriptorArchivingScheduledToProducer.js";
 import { handleEserviceDescriptorArchivingScheduledToConsumer } from "./handleEserviceDescriptorArchivingScheduledToConsumer.js";
@@ -20,7 +22,6 @@ import { handleEserviceDescriptorArchivingCompletedToConsumer } from "./handleEs
 import { handleEserviceArchivingCompletedToProducer } from "./handleEserviceArchivingCompletedToProducer.js";
 import { handleEserviceArchivingCompletedToConsumer } from "./handleEserviceArchivingCompletedToConsumer.js";
 import { handleEserviceDescriptorArchivedToProducer } from "./handleEserviceDescriptorArchivedToProducer.js";
-import { handleEserviceDescriptorArchivedToConsumer } from "./handleEserviceDescriptorArchivedToConsumer.js";
 import { handleEserviceArchivingCanceledToConsumer } from "./handleEserviceArchivingCanceledToConsumer.js";
 import { handleEserviceDescriptorArchivingCanceledToConsumer } from "./handleEserviceDescriptorArchivingCanceledToConsumer.js";
 
@@ -53,27 +54,51 @@ export async function handleEServiceEvent(
     )
     .with(
       { type: "EServiceDescriptorActivated" },
-      ({ data: { eservice, descriptorId } }) =>
-        handleEserviceDescriptorActivated({
-          eserviceV2Msg: eservice,
-          descriptorId,
-          logger,
-          readModelService,
-          templateService,
-          correlationId,
-        })
+      async ({ data: { eservice, descriptorId } }) => {
+        const [prod, cons] = await Promise.all([
+          handleEserviceDescriptorActivatedToProducer({
+            eserviceV2Msg: eservice,
+            descriptorId,
+            logger,
+            readModelService,
+            templateService,
+            correlationId,
+          }),
+          handleEserviceDescriptorActivatedToConsumer({
+            eserviceV2Msg: eservice,
+            descriptorId,
+            logger,
+            readModelService,
+            templateService,
+            correlationId,
+          }),
+        ]);
+        return [...prod, ...cons];
+      }
     )
     .with(
       { type: "EServiceDescriptorSuspended" },
-      ({ data: { eservice, descriptorId } }) =>
-        handleEserviceDescriptorSuspended({
-          eserviceV2Msg: eservice,
-          descriptorId,
-          logger,
-          readModelService,
-          templateService,
-          correlationId,
-        })
+      async ({ data: { eservice, descriptorId } }) => {
+        const [prod, cons] = await Promise.all([
+          handleEserviceDescriptorSuspendedToProducer({
+            eserviceV2Msg: eservice,
+            descriptorId,
+            logger,
+            readModelService,
+            templateService,
+            correlationId,
+          }),
+          handleEserviceDescriptorSuspendedToConsumer({
+            eserviceV2Msg: eservice,
+            descriptorId,
+            logger,
+            readModelService,
+            templateService,
+            correlationId,
+          }),
+        ]);
+        return [...prod, ...cons];
+      }
     )
     .with(
       { type: "EServiceDescriptorSubmittedByDelegate" },
@@ -225,27 +250,15 @@ export async function handleEServiceEvent(
     )
     .with(
       { type: "EServiceDescriptorArchived" },
-      async ({ data: { eservice, descriptorId } }) => {
-        const [prod, cons] = await Promise.all([
-          handleEserviceDescriptorArchivedToProducer({
-            eserviceV2Msg: eservice,
-            descriptorId,
-            logger,
-            readModelService,
-            templateService,
-            correlationId,
-          }),
-          handleEserviceDescriptorArchivedToConsumer({
-            eserviceV2Msg: eservice,
-            descriptorId,
-            logger,
-            readModelService,
-            templateService,
-            correlationId,
-          }),
-        ]);
-        return [...prod, ...cons];
-      }
+      ({ data: { eservice, descriptorId } }) =>
+        handleEserviceDescriptorArchivedToProducer({
+          eserviceV2Msg: eservice,
+          descriptorId,
+          logger,
+          readModelService,
+          templateService,
+          correlationId,
+        })
     )
     .with(
       { type: "EServiceArchivingCanceled" },
