+++ packages/email-notification-dispatcher/test/handleEserviceDescriptorActivatedToConsumer.test.ts
@@ -25,7 +25,7 @@ import {
   descriptorNotFound,
   tenantNotFound,
 } from "pagopa-interop-notification-commons";
-import { handleEserviceDescriptorActivated } from "../src/handlers/eservices/handleEserviceDescriptorActivated.js";
+import { handleEserviceDescriptorActivatedToConsumer } from "../src/handlers/eservices/handleEserviceDescriptorActivatedToConsumer.js";
 import {
   addOneAgreement,
   addOneEService,
@@ -35,7 +35,7 @@ import {
   templateService,
 } from "./utils.js";
 
-describe("handleEserviceDescriptorActivated", async () => {
+describe("handleEserviceDescriptorActivatedToConsumer", async () => {
   const producerId = generateId<TenantId>();
   const consumerIds = [generateId<TenantId>(), generateId<TenantId>()];
   const eserviceId = generateId<EServiceId>();
@@ -79,7 +79,7 @@ describe("handleEserviceDescriptorActivated", async () => {
 
   it("should throw missingKafkaMessageDataError when eservice is undefined", async () => {
     await expect(() =>
-      handleEserviceDescriptorActivated({
+      handleEserviceDescriptorActivatedToConsumer({
         eserviceV2Msg: undefined,
         descriptorId: descriptor.id,
         logger,
@@ -102,7 +102,7 @@ describe("handleEserviceDescriptorActivated", async () => {
     };
 
     await expect(() =>
-      handleEserviceDescriptorActivated({
+      handleEserviceDescriptorActivatedToConsumer({
         eserviceV2Msg: toEServiceV2(eserviceWithUnknownProducer),
         descriptorId: descriptor.id,
         logger,
@@ -117,7 +117,7 @@ describe("handleEserviceDescriptorActivated", async () => {
     const nonExistentDescriptorId = generateId<DescriptorId>();
 
     await expect(() =>
-      handleEserviceDescriptorActivated({
+      handleEserviceDescriptorActivatedToConsumer({
         eserviceV2Msg: toEServiceV2(eservice),
         descriptorId: nonExistentDescriptorId,
         logger,
@@ -129,7 +129,7 @@ describe("handleEserviceDescriptorActivated", async () => {
   });
 
   it("should return empty array if no consumer is present for the eservice", async () => {
-    const messages = await handleEserviceDescriptorActivated({
+    const messages = await handleEserviceDescriptorActivatedToConsumer({
       eserviceV2Msg: toEServiceV2(eservice),
       descriptorId: descriptor.id,
       logger,
@@ -153,7 +153,7 @@ describe("handleEserviceDescriptorActivated", async () => {
     await addOneAgreement(agreements[0]);
     await addOneAgreement(agreements[1]);
 
-    const messages = await handleEserviceDescriptorActivated({
+    const messages = await handleEserviceDescriptorActivatedToConsumer({
       eserviceV2Msg: toEServiceV2(eservice),
       descriptorId: descriptor.id,
       logger,
@@ -214,7 +214,7 @@ describe("handleEserviceDescriptorActivated", async () => {
     await addOneAgreement(agreements[0]);
     await addOneAgreement(agreements[1]);
 
-    const messages = await handleEserviceDescriptorActivated({
+    const messages = await handleEserviceDescriptorActivatedToConsumer({
       eserviceV2Msg: toEServiceV2(eservice),
       descriptorId: descriptor.id,
       logger,
@@ -258,7 +258,7 @@ describe("handleEserviceDescriptorActivated", async () => {
     };
     await addOneAgreement(agreement);
 
-    const messages = await handleEserviceDescriptorActivated({
+    const messages = await handleEserviceDescriptorActivatedToConsumer({
       eserviceV2Msg: toEServiceV2(eservice),
       descriptorId: descriptor.id,
       logger,
@@ -294,7 +294,7 @@ describe("handleEserviceDescriptorActivated", async () => {
     await addOneAgreement(agreements[0]);
     await addOneAgreement(agreements[1]);
 
-    const messages = await handleEserviceDescriptorActivated({
+    const messages = await handleEserviceDescriptorActivatedToConsumer({
       eserviceV2Msg: toEServiceV2(eservice),
       descriptorId: descriptor.id,
       logger,
