+++ packages/email-notification-dispatcher/test/handleEserviceDescriptorActivatedToProducer.test.ts
@@ -1,0 +1,131 @@
+/* eslint-disable functional/immutable-data */
+import {
+  getMockContext,
+  getMockDescriptor,
+  getMockDescriptorPublished,
+  getMockEService,
+  getMockTenant,
+} from "pagopa-interop-commons-test";
+import { authRole } from "pagopa-interop-commons";
+import {
+  CorrelationId,
+  Descriptor,
+  descriptorState,
+  EService,
+  EServiceId,
+  archivingScope,
+  generateId,
+  missingKafkaMessageDataError,
+  TenantId,
+  toEServiceV2,
+} from "pagopa-interop-models";
+import { beforeEach, describe, expect, it, vi } from "vitest";
+import { handleEserviceDescriptorActivatedToProducer } from "../src/handlers/eservices/handleEserviceDescriptorActivatedToProducer.js";
+import {
+  addOneEService,
+  addOneTenant,
+  getMockUser,
+  readModelService,
+  templateService,
+} from "./utils.js";
+
+describe("handleEserviceDescriptorActivatedToProducer", () => {
+  const producerId = generateId<TenantId>();
+  const producerTenant = { ...getMockTenant(producerId), name: "Producer T" };
+
+  const archivingDescriptor: Descriptor = {
+    ...getMockDescriptor(descriptorState.archiving),
+    archivingSchedule: {
+      archivableOn: new Date("2026-12-31T00:00:00.000Z"),
+      startedAt: new Date("2026-05-14T00:00:00.000Z"),
+      scope: archivingScope.descriptor,
+    },
+  };
+  const eservice: EService = {
+    ...getMockEService(),
+    name: "Test E-service",
+    producerId,
+    descriptors: [archivingDescriptor],
+  };
+
+  const users = [
+    getMockUser(producerTenant.id),
+    getMockUser(producerTenant.id),
+  ];
+
+  const { logger } = getMockContext({});
+
+  const expectedMessageBody =
+    /<p>La versione \d+ dell'e-service "[^"]+" è di nuovo attiva. I fruitori potranno nuovamente scambiare dati con questa versione.<\/p>[\s\n]+<p>La versione è in fase di archiviazione e sarà archiviata definitivamente il giorno \d{2}\/\d{2}\/\d{4}\./;
+
+  beforeEach(async () => {
+    await addOneEService(eservice);
+    await addOneTenant(producerTenant);
+    readModelService.getTenantUsersWithNotificationEnabled = vi
+      .fn()
+      .mockImplementation((tenantIds: TenantId[]) =>
+        users
+          .filter((u) => tenantIds.includes(u.tenantId))
+          .map((u) => ({
+            userId: u.id,
+            tenantId: u.tenantId,
+            userRoles: [authRole.ADMIN_ROLE],
+          }))
+      );
+  });
+
+  it("throws missingKafkaMessageDataError when eservice is undefined", async () => {
+    await expect(() =>
+      handleEserviceDescriptorActivatedToProducer({
+        eserviceV2Msg: undefined,
+        descriptorId: archivingDescriptor.id,
+        logger,
+        templateService,
+        readModelService,
+        correlationId: generateId<CorrelationId>(),
+      })
+    ).rejects.toThrow(
+      missingKafkaMessageDataError("eservice", "EServiceDescriptorActivated")
+    );
+  });
+
+  it("emits one email per producer user with the expected subject and body when archivingSchedule is present on the descriptor", async () => {
+    const messages = await handleEserviceDescriptorActivatedToProducer({
+      eserviceV2Msg: toEServiceV2(eservice),
+      descriptorId: archivingDescriptor.id,
+      logger,
+      templateService,
+      readModelService,
+      correlationId: generateId<CorrelationId>(),
+    });
+    expect(messages).toHaveLength(users.length);
+    expect(messages[0].email.subject).toBe(
+      `Una versione di "${eservice.name}" è stata riattivata`
+    );
+    expect(messages[0].email.body).toMatch(expectedMessageBody);
+  });
+
+  it("emits no email per producer user when archivingSchedule is absent on the descriptor (routine auto-archive)", async () => {
+    const routineDescriptor: Descriptor = {
+      ...getMockDescriptorPublished(),
+    };
+    const routineEservice: EService = {
+      ...getMockEService(),
+      id: generateId<EServiceId>(),
+      producerId,
+      descriptors: [routineDescriptor],
+    };
+    await addOneEService(routineEservice);
+    await addOneTenant({ ...getMockTenant(producerId), name: "Producer T" });
+
+    const messages = await handleEserviceDescriptorActivatedToProducer({
+      eserviceV2Msg: toEServiceV2(routineEservice),
+      descriptorId: routineDescriptor.id,
+      logger,
+      templateService,
+      readModelService,
+      correlationId: generateId<CorrelationId>(),
+    });
+    expect(messages).toHaveLength(0);
+  });
+});
