+++ packages/notification-commons/src/templates/email/resources/templates/eservice-archiving-descriptor-activated-to-consumer-mail.html
