+++ packages/notification-commons/src/templates/email/resources/templates/eservice-archiving-descriptor-suspended-to-consumer-mail.html
