+++ packages/notification-commons/src/templates/email/resources/templates/eservice-archiving-descriptor-suspended-to-producer-mail.html
@@ -23,6 +23,7 @@
         color: #17324d;
       "
     >
+      <!-- H4 Desktop (from MUI Italia)-->
       <h1 style="font-size: 32px; line-height: 40px; margin: 0">{{ title }}</h1>
     </div>
   </td>
@@ -74,7 +75,9 @@
         color: #17324d;
       "
     >
-      <p>L'ente erogatore <strong>{{ producerName }}</strong> ha archiviato in anticipo la versione {{ eserviceVersion }} dell'e-service "{{ eserviceName }}".</p>
+      <!-- Body 1 Desktop (from MUI Italia)-->
+       <p>La versione {{ eserviceVersion }} dell'e-service "{{ eserviceName }}" è al momento sospesa. I fruitori non potranno più scambiare dati con questa versione.</p>
+       <p>La versione è in fase di archiviazione e sarà archiviata definitivamente il giorno {{ archivableOn }}.</p>
     </div>
   </td>
 </tr>
